+++ src/ui/event/mod.rs
@@ -4,10 +4,12 @@
 //
 
 mod event; // Event
+mod id; // DeviceId, WindowId
 mod key; // EventKey[Ffi], Key[Ffi|Media|Mod|Mods|Pad|State]
 mod kind; // EventKind, EventTag
 mod pointer; // Event[Button[State]|Mouse|Pointer[Type]|Wheel]
 mod queue; // EventQueue
+mod target; // EventTarget
 mod time; // EventTimestamp
 mod window; // EventWindow
 
@@ -18,10 +20,12 @@ crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             event::*,
+            id::*,
             key::*,
             kind::*,
             pointer::*,
             queue::*,
+            target::*,
             time::*,
             window::*,
         };